silence a couple minor warnings

+++ Library/Sources/Stroika/Foundation/IO/Network/InternetAddress.cpp
@@ -160,7 +160,7 @@ namespace Stroika::Foundation::IO::Network {
         for (uint8_t b : As<vector<uint8_t>> ()) {
             // We could logically have numbered bits either way, but since this returns high order byte first only makes
             // sense to return high order bit first
-            for (size_t i = 0; i < 8; ++i) {
+            for (unsigned int i = 0; i < 8; ++i) {
                 result.push_back (BitSubstring (b, 7 - i, 7 - i + 1));
             }
         }
